pyproject.toml
[project]
name = "resskills"
version = "0.1.0"
description = "Virtual AI research team for Claude Code"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = ["jinja2>=3.1", "pyyaml>=6.0"]

[project.scripts]
resskills-gen = "resskills.scripts.gen_skills:main"
resskills-config = "resskills.scripts.config:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["resskills"]
force-single-line = false
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
